Merry Christmas MaxiCM(1/2)
This Christmas we want to wish our customers in a special way best wishes
for 2016, and as a happy Christmas, that's why the team decided
to prepare this MaxiCM commit to users who are installing their OTA or
installed by MaxiCM first time with a beautiful surprise.
We wish you a Merry Christmas to all of you,
and that your 2016 is full of surprises, and believe us,
we have many more surprises in store for you.
Thank you for supporting our crazy all this 2015 project,
and more things will come this 2016.

Greetings, MaxiCM users.

Signed-off-by: Krizthian <cristianortizmorales2013@gmail.com>