Merry Christmas MaxiCM(2/2)
Normally developers MaxiCM work hard in the source code to take each week
a build, but also people who want to port
to MaxiCM your device is also quite struggling,
that is why we also want to surprise them, so if you read this and you're
done building MaxiCM, go take a break and take a delicious cup
of hot chocolate. We also take a break or maybe a little
more work, but the important thing is that
at this time you spend with family, so take a break. Working on Android
takes time, time that you could
spend your family. So leave it aside for a while and go enjoy.

And do not forget that 2016
will soon be so still resting after Christmas because come 2016.
We want you to have a Merry Christmas and a Happy New Year 2016.

Greetings, MaxiCM Mantainers.

Signed-off-by: Krizthian <cristianortizmorales2013@gmail.com>